requirements.txt
streamlit
joblib
pandas
numpy
scikit-learn
concurrent-futures
propy3
openpyxl